+++ packages/connect-multichain/src/domain/multichain/types.ts
@@ -1,5 +1,9 @@
 import type { SessionRequest } from '@metamask/mobile-wallet-protocol-core';
-import type { Transport, TransportRequest, TransportResponse } from '@metamask/multichain-api-client';
+import type {
+  Transport,
+  TransportRequest,
+  TransportResponse,
+} from '@metamask/multichain-api-client';
 import type { CaipAccountId } from '@metamask/utils';
 
 import type { MultichainCore } from '.';
@@ -93,9 +97,16 @@ export type ExtendedTransport = Omit<Transport, 'connect'> & {
   connect: (props?: {
     scopes: Scope[];
     caipAccountIds: CaipAccountId[];
+    forceRequest?: boolean;
   }) => Promise<void>;
 
-  sendEip1193Message: <TRequest extends TransportRequest, TResponse extends TransportResponse>(request: TRequest, options?: {
-    timeout?: number;
-}) => Promise<TResponse>;
+  sendEip1193Message: <
+    TRequest extends TransportRequest,
+    TResponse extends TransportResponse,
+  >(
+    request: TRequest,
+    options?: {
+      timeout?: number;
+    },
+  ) => Promise<TResponse>;
 };