+++ packages/connect-multichain/src/multichain/index.ts
@@ -515,6 +515,7 @@ export class MultichainSDK extends MultichainCore {
   async connect(
     scopes: Scope[],
     caipAccountIds: CaipAccountId[],
+    forceRequest?: boolean,
   ): Promise<void> {
     const { ui } = this.options;
     const platformType = getPlatformType();
@@ -527,13 +528,15 @@ export class MultichainSDK extends MultichainCore {
 
     if (this.__transport?.isConnected() && !secure) {
       return this.handleConnection(
-        this.__transport.connect({ scopes, caipAccountIds }).then(() => {
-          if (this.__transport instanceof MWPTransport) {
-            return this.storage.setTransport(TransportType.MPW);
-          } else {
-            return this.storage.setTransport(TransportType.Browser);
-          }
-        }),
+        this.__transport
+          .connect({ scopes, caipAccountIds, forceRequest })
+          .then(() => {
+            if (this.__transport instanceof MWPTransport) {
+              return this.storage.setTransport(TransportType.MPW);
+            } else {
+              return this.storage.setTransport(TransportType.Browser);
+            }
+          }),
       );
     }
 
@@ -541,7 +544,7 @@ export class MultichainSDK extends MultichainCore {
     if (platformType === PlatformType.MetaMaskMobileWebview) {
       const defaultTransport = await this.setupDefaultTransport();
       return this.handleConnection(
-        defaultTransport.connect({ scopes, caipAccountIds }),
+        defaultTransport.connect({ scopes, caipAccountIds, forceRequest }),
       );
     }
 
@@ -550,7 +553,7 @@ export class MultichainSDK extends MultichainCore {
       const defaultTransport = await this.setupDefaultTransport();
       // Web transport has no initial payload
       return this.handleConnection(
-        defaultTransport.connect({ scopes, caipAccountIds }),
+        defaultTransport.connect({ scopes, caipAccountIds, forceRequest }),
       );
     }
 
@@ -619,7 +622,11 @@ export class MultichainSDK extends MultichainCore {
         if (mobile?.preferredOpenLink) {
           mobile.preferredOpenLink(METAMASK_DEEPLINK_BASE, '_self');
         } else {
-          openDeeplink(this.options, METAMASK_DEEPLINK_BASE, METAMASK_CONNECT_BASE_URL);
+          openDeeplink(
+            this.options,
+            METAMASK_DEEPLINK_BASE,
+            METAMASK_CONNECT_BASE_URL,
+          );
         }
       }, 10); // small delay to ensure the message encryption and dispatch completes
     }